feat(#47): db 모듈 분리
